refactor: removes duplicate code by creating a generic model
